Add Configuration. Fix checkstyle violation
